Removed extra files from prev cmt
